Decouple results store and artefacts store
This commit decouples the functionality of choosing and where to store
the results of the validations, and the functionality that enables
passing an artefacts store URI to validations. Substantively the two
arguments `results_path` and `artefacts_store_directory`, which replaces
`store_directory` and `pass_artefact_directory`, don't have any
interaction with each other.

This change also enables the user to set the name of the validation
results file if they wish to.